CSC110 Project/requirements.txt
beautifulsoup4
requests
numpy
Pillow